move observers models to reusable package
